This is Second Commit. (Notification and other small updates)

+++ app/src/main/java/com/example/weatherapp/Wind.kt
@@ -2,6 +2,6 @@ package com.example.weatherapp
 
 data class Wind(
     val deg: Int,
-    val gust: Int,
+    val gust: Double,
     val speed: Double
 )
